Refactor: Компонент MovieCard переименован в SmallMovieCard для большей индентичности с разметкой
