+++ src/components/small-movie-card/small-movie-card.jsx
